update revisi usia

+++ app/src/main/java/com/sipemandu/sipemandu/Room/Model/DataAnak.java
@@ -14,11 +14,11 @@ public class DataAnak {
     private String namaAnak;
     private String jenisKelamin;
     private String tanggalLahir;
-    private int beratBadan;
-    private int tinggiBadan;
+    private Double beratBadan;
+    private Double tinggiBadan;
     private String asiEksklusif;
 
-    public DataAnak(String nikAnak, String namaAnak, String jenisKelamin, String tanggalLahir, int beratBadan, int tinggiBadan, String asiEksklusif) {
+    public DataAnak(String nikAnak, String namaAnak, String jenisKelamin, String tanggalLahir, Double beratBadan, Double tinggiBadan, String asiEksklusif) {
         this.nikAnak = nikAnak;
         this.namaAnak = namaAnak;
         this.jenisKelamin = jenisKelamin;
@@ -53,11 +53,11 @@ public class DataAnak {
         return tanggalLahir;
     }
 
-    public int getBeratBadan() {
+    public Double getBeratBadan() {
         return beratBadan;
     }
 
-    public int getTinggiBadan() {
+    public Double getTinggiBadan() {
         return tinggiBadan;
     }
